requirements.txt
absl-py==0.7.1
alt-model-checkpoint==1.0.1
astor==0.8.0
bleach==1.5.0
cycler==0.10.0
data==0.4
decorator==4.4.0
funcsigs==1.0.2
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
html5lib==0.9999999
imageio==2.5.0
Keras==2.1.6
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
mock==3.0.5
networkx==2.3
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tensorboard==1.9.0
tensorboardX==1.7
tensorflow-gpu==1.9.0
termcolor==1.1.0
Werkzeug==0.15.4
